# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/staff/reports/popular": {
      "get": {
        "operationId": "ReportsController_popular",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PopularRowDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "reports"
        ]
      }
    },
    "/api/staff/reports/volume": {
      "get": {
        "operationId": "ReportsController_volume",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "interval",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "day",
                "month"
              ],
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VolumeRowDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "reports"
        ]
      }
    },
    "/api/staff/reports/overdue": {
      "get": {
        "operationId": "ReportsController_overdue",
        "parameters": [
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OverdueRowDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "reports"
        ]
      }
    },
    "/api/staff/reports/fines": {
      "get": {
        "operationId": "ReportsController_fines",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "interval",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "day",
                "month"
              ],
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinesReportDto"
                }
              }
            }
          }
        },
        "tags": [
          "reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "BiblioApp API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PopularRowDto": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "authors": {
            "type": "string"
          },
          "loans": {
            "type": "number"
          }
        },
        "required": [
          "bookId",
          "title",
          "authors",
          "loans"
        ]
      },
      "VolumeRowDto": {
        "type": "object",
        "properties": {
          "period": {
            "type": "string"
          },
          "loans": {
            "type": "number"
          },
          "returns": {
            "type": "number"
          }
        },
        "required": [
          "period",
          "loans",
          "returns"
        ]
      },
      "OverdueRowDto": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "memberNumber": {
            "type": "string"
          },
          "memberName": {
            "type": "string"
          },
          "dueAt": {
            "type": "string"
          },
          "daysLate": {
            "type": "number"
          },
          "fineCents": {
            "type": "number"
          }
        },
        "required": [
          "loanId",
          "title",
          "memberNumber",
          "memberName",
          "dueAt",
          "daysLate",
          "fineCents"
        ]
      },
      "FinePeriodDto": {
        "type": "object",
        "properties": {
          "period": {
            "type": "string"
          },
          "issuedCents": {
            "type": "number"
          },
          "collectedCents": {
            "type": "number"
          }
        },
        "required": [
          "period",
          "issuedCents",
          "collectedCents"
        ]
      },
      "FinesReportDto": {
        "type": "object",
        "properties": {
          "issuedCents": {
            "type": "number"
          },
          "collectedCents": {
            "type": "number"
          },
          "waivedCents": {
            "type": "number"
          },
          "outstandingCents": {
            "type": "number"
          },
          "byPeriod": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FinePeriodDto"
            }
          }
        },
        "required": [
          "issuedCents",
          "collectedCents",
          "waivedCents",
          "outstandingCents",
          "byPeriod"
        ]
      },
...
    }
  }
}